BS-13234 generate input contract from selected business data field
refactoring
