+++ bundles/plugins/org.bonitasoft.studio.contract/src/org/bonitasoft/studio/contract/core/mapping/FieldToContractInputMapping.java
@@ -14,27 +14,18 @@
  */
 package org.bonitasoft.studio.contract.core.mapping;
 
-import static com.google.common.base.Predicates.and;
-import static com.google.common.base.Predicates.instanceOf;
-
 import java.util.ArrayList;
 import java.util.List;
 
 import org.bonitasoft.engine.bdm.model.field.Field;
-import org.bonitasoft.engine.bdm.model.field.FieldType;
-import org.bonitasoft.engine.bdm.model.field.RelationField;
-import org.bonitasoft.engine.bdm.model.field.RelationField.Type;
-import org.bonitasoft.engine.bdm.model.field.SimpleField;
 import org.bonitasoft.studio.model.process.ContractInput;
 import org.bonitasoft.studio.model.process.ContractInputType;
 import org.bonitasoft.studio.model.process.ProcessFactory;
 
-import com.google.common.base.Predicate;
-
 /**
  * @author aurelie
  */
-public class FieldToContractInputMapping {
+public abstract class FieldToContractInputMapping {
 
     private final Field field;
 
@@ -64,70 +55,13 @@ public class FieldToContractInputMapping {
         return parent;
     }
 
-    /**
-     * @return
-     */
     public ContractInput toContractInput() {
         final ContractInput contractInput = ProcessFactory.eINSTANCE.createContractInput();
         contractInput.setName(field.getName());
         contractInput.setType(toContractInputType());
-        if (shouldAddChildInput()) {
-            for (final FieldToContractInputMapping child : getChildren()) {
-                contractInput.getInputs().add(child.toContractInput());
-            }
-        }
         return contractInput;
     }
 
-    private boolean shouldAddChildInput() {
-        return and(instanceOf(RelationField.class), withType(Type.COMPOSITION)).apply(field);
-    }
+    protected abstract ContractInputType toContractInputType();
 
-    private Predicate<Field> withType(final Type relationType) {
-        return new Predicate<Field>() {
-
-            @Override
-            public boolean apply(final Field input) {
-                return ((RelationField) input).getType() == relationType;
-            }
-        };
-    }
-
-    private ContractInputType toContractInputType() {
-        if (field instanceof SimpleField) {
-            final FieldType fieldType = ((SimpleField) field).getType();
-            switch (fieldType) {
-                case STRING:
-                    return ContractInputType.TEXT;
-                case TEXT:
-                    return ContractInputType.TEXT;
-                case INTEGER:
-                    return ContractInputType.INTEGER;
-                case LONG:
-                    return ContractInputType.INTEGER;
-                case DOUBLE:
-                    return ContractInputType.DECIMAL;
-                case FLOAT:
-                    return ContractInputType.DECIMAL;
-                case BOOLEAN:
-                    return ContractInputType.BOOLEAN;
-                case DATE:
-                    return ContractInputType.DATE;
-                default:
-                    throw new IllegalStateException(String.format("Failed to convert field type %s to contract input type", fieldType));
-            }
-        }
-        if (field instanceof RelationField) {
-            final Type relationType = ((RelationField) field).getType();
-            switch (relationType) {
-                case COMPOSITION:
-                    return ContractInputType.COMPLEX;
-                case AGGREGATION:
-                    return ContractInputType.TEXT;
-                default:
-                    throw new IllegalStateException(String.format("Failed to convert field type %s to contract input type", relationType));
-            }
-        }
-        throw new IllegalStateException("Unsupported field type.");
-    }
 }
